core: ensure newly-created master interfaces are ready for activation
Reverts part of

2226a00cc2400a4150a83c119c8bfcc06ff2e988
core: add a "default-unmanaged" setting for devices

Newly-created master interfaces are in the UNAVAILABLE state, but if
they were created in response to a slave connection being activated,
the master must be activated immediately too.  But a device cannot
be activated unless it's in the DISCONNECTED state, so restore that
state change.