bluez: merge file nm-bluez5-device.c into nm-bluez-device.c
The files nm-bluez-device.c and nm-bluez5-device.c are similar.
Merge them together into one file and handle the differences
using #if directives.

This commit does not actually change any functionality. All it does, is
merging the files together and separating the differences with #if#else.

The next commit will further add common functionality, so this
reduces code duplication for now.

Signed-off-by: Thomas Haller <thaller@redhat.com>