trivial: quiet log message about failing to determine virtual interface name
In the case of autoconnect VLANs or IB partitions, if the parent interface
hasn't been detected yet at startup, then the get_virtual_interface_name()
won't be able to find the parent yet.  That's normal, and when the parent
is found, system_create_virtual_device() will be run again and the parent
will be found, and the autoconnect VLAN/IB partition will be created.

But we shouldn't warn that the parent can't be found when that might be
a normal occurance.