man: NetworkManager.conf.5 fixes and updates
The [main] section is not mandatory.

Clarify in several places that the keyfile plugin is always used for
fallback, and that the [keyfile] section is normally only used if you
aren't using any other plugin.

Fix some erroneous references to "keyfile" and "ifdown" in the
ifupdown section.

Update the ifcfg-rh docs to list all currently-supported connection
types.

Swap the order of ifcfg-suse and ifupdown to make them alphabetical.
(Note that ifnet is currently missing.)

https://bugzilla.gnome.org/show_bug.cgi?id=720841