contrib/nm-live-vm: add rootfstype=ramfs kernel parameter to run.sh
Kernel changed default filesystem for roots from ramfs to tmpfs.
See http://lwn.net/Articles/559176/ for more information.
The change caused our initramfs to be mounted with size= parameter that equals
to the actual size of unpacked initramfs. That's why mounted / was full and no
space was available (without manual remounting: mount -o remount,size=100% /).

So we explicitly require usage of ramfs to have all RAM of the virtual machine
available for /.

Signed-off-by: Jiří Klimeš <jklimes@redhat.com>